third_party/blink/web_tests/platform/generic/fast/events/touch/touch-inside-iframe-expected.txt

Test iframes receive touches correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS touch.pageX is 50
PASS touch.pageY is 50
PASS successfullyParsed is true

TEST COMPLETE
